Refs #34548 - move HostCollectionsCard.js
